requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-openai
openai
faiss-cpu
pydantic
numpy
pandas
python-dotenv
litellm
tabulate